requirements.txt
pyodbc==4.0.30
Flask==1.1.2
numpy==1.18.2
pandas==1.0.1
requests==2.23.0
jsonschema==3.2.0
azure-identity==1.3.1
azure-keyvault-secrets
msrestazure==0.5.0
pyarrow==0.16.0
flask-restful
msrest

















